Build and Nuke plugins revamp
The whole build system is now using CMake
Additional work to build with Conan and Docker to provide a as-binary-compatible-as-possible-on-Linux plugin suite
Check the INSTALL notes on Linux to find out how to build with Docker

Big clean up of the Nuke plugins
The plugin generate hashes of the input metadata to reduce the whole depacking/decoding overhead as much as possible
The input patterns (names AND lpes) are also hashed to prevent rebuilding the automating all the time, and eventually
the active names/lpes are also hashed to cache the result of all regexes/lpes evaluation which can be long when
dealing with massive scenes. All multithread safe.
The downside is we've switched to c++11, which prevents us from compiling plugins for Nuke9 (since gcc4.1 and vs2010
won't compile this.) Porting back to old c++ is not really considered at this point.